src/main/java/com/transactionroutine/app/service/impl/AccountServiceImpl.java
package com.transactionroutine.app.service.impl;


import com.transactionroutine.app.exception.NotFoundException;
import com.transactionroutine.app.model.Account;
import com.transactionroutine.app.repository.AccountRepository;
import com.transactionroutine.app.service.AccountService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class AccountServiceImpl implements AccountService {

    private final AccountRepository accountRepository;

    @Autowired
    public AccountServiceImpl(AccountRepository accountRepository) {
        this.accountRepository = accountRepository;
    }

    @Override
    public Account createAccount(String documentNumber) {
        Account account = new Account();
        account.setDocumentNumber(documentNumber);
        return accountRepository.save(account);
    }

    @Override
    public Account getAccountById(Long id) {
        return accountRepository.findById(id)
                .orElseThrow(() -> new NotFoundException("Account not found with id: " + id));
    }
}
